Black Jack: refactor
Update the 'higher_card' function so that it would match the case
faster. I've pushed the '==' case further below since it is the least
possible case.

Simplify the 'value_of_ace' function logic. It doesn't need a match
statement since the logic could be much straightforward. If there is an ace
already in hand, the second can't be 11 since the sum would go over 21 and
the player goes 'bust'. On the contrary, if there is no ace in the hand but
the card values sum up to more than 10 whatsoever, the new ace must be 1 to
prevent the player from going 'bust' again. Otherwise, we can simply return
11 since there is no ace in the hand and the sum would not go over 21.

Add additional documentation about "splitting pairs" and "is_blackjack"
functions.