Strings: refactor, use string methods
Until now, the strings.py file used regex to substitute trailing
punctuations. However, the same could be done using built-in string methods
which is more readable.

Also, I've updated the logic in 'make_word_groups' so that there is no need
for an additional string concatenation.

Removed the extra constant which I forgot to delete in the previous
iteration.