Pangram: README, the set method
There are three approaches shared on the exercism website. I implemented
the 'all' approach by myself in 15 minutes (and thought that was cool, then
I was surprised that the website contained the exact same code that I
wrote!). But then, the 'set' approach looked much more interesting and
performant.

Also, update the README file.