Darts: prevent human error, add comments
In the previous iteration, the 'circles' tuple had to be in the ascending
order of the circle radius or the logic would fail. Since the developers
might not be aware of that (and skip the documentation) I simply sorted the
'circles' by their radius.

Also, add comments so that the developers know 'why' the list is sorted.