Perfect Numbers: refactor after code review
I tried using list comprehensions, mapping, etc. Which complicated the
logic. The same task could be done by simply using a for loop instead
of sticking to the list comprehensions which rises the need to unpack
the tuples later.

At first I was skeptical about updating the "set" in every
iteration. But the performance is still virtually identical to the
previous commit.

This one is easier to read, and has the same performance. So let's go
with it.