ISBN Verifier: refactor, use simpler logic
Checking for non-digits is easier and more readable when using "all"
instead of checking for characters one by one.

Making a list of the initial string is smarter and cleaner. Replacing
the check digit is also easier this way.

Using a simple "sum" is more succint than a for loop.